# server/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: verification-service
  version: 1.0.0
paths:
  /:
    get:
      summary: Root
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /utc-comparison:
    post:
      summary: Compare UTC outputs for each species in a given model file. You may
        pass either a model file or OMEX archive file.
      operationId: utc-comparison
      parameters:
      - name: simulators
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - amici
          - copasi
          - tellurium
          title: Simulators
      - name: include_outputs
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Outputs
      - name: comparison_id
        in: query
        required: false
        schema:
          type: string
          title: Comparison Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_utc-comparison'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UtcComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /biosimulators-utc-comparison:
    post:
      summary: Compare UTC outputs from Biosimulators for a model from a given archive.
      operationId: biosimulators-utc-comparison
      parameters:
      - name: simulators
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          description: Simulators to include in the comparison.
          default:
          - amici
          - copasi
          - tellurium
          title: Simulators
        description: Simulators to include in the comparison.
      - name: include_outputs
        in: query
        required: false
        schema:
          type: boolean
          description: Whether to include the output data on which the comparison
            is based.
          default: true
          title: Include Outputs
        description: Whether to include the output data on which the comparison is
          based.
      - name: comparison_id
        in: query
        required: false
        schema:
          type: string
          description: Descriptive identifier for this comparison.
          title: Comparison Id
        description: Descriptive identifier for this comparison.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_biosimulators-utc-comparison'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UtcComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /biosimulators-utc-species-comparison:
    post:
      summary: Compare UTC outputs from Biosimulators for a given species name
      operationId: biosimulators_utc_species_comparison_biosimulators_utc_species_comparison_post
      parameters:
      - name: species_id
        in: query
        required: true
        schema:
          type: string
          title: Species Id
      - name: simulators
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          default:
          - amici
          - copasi
          - tellurium
          title: Simulators
      - name: include_outputs
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Outputs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_biosimulators_utc_species_comparison_biosimulators_utc_species_comparison_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UtcSpeciesComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_biosimulators-utc-comparison:
      properties:
        uploaded_file:
          type: string
          format: binary
          title: Uploaded File
          description: OMEX/COMBINE Archive File.
        ground_truth_report:
          type: string
          format: binary
          title: Ground Truth Report
          description: reports.h5 file defining the so-called ground-truth to be included
            in the comparison.
      type: object
      required:
      - uploaded_file
      title: Body_biosimulators-utc-comparison
    Body_biosimulators_utc_species_comparison_biosimulators_utc_species_comparison_post:
      properties:
        uploaded_file:
          type: string
          format: binary
          title: Uploaded File
      type: object
      required:
      - uploaded_file
      title: Body_biosimulators_utc_species_comparison_biosimulators_utc_species_comparison_post
    Body_utc-comparison:
      properties:
        uploaded_file:
          type: string
          format: binary
          title: Uploaded File
      type: object
      required:
      - uploaded_file
      title: Body_utc-comparison
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    UtcComparison:
      properties:
        results:
          items:
            $ref: '#/components/schemas/UtcSpeciesComparison'
          type: array
          title: Results
        id:
          type: string
          title: Id
        simulators:
          items:
            type: string
          type: array
          title: Simulators
      type: object
      required:
      - results
      - id
      - simulators
      title: UtcComparison
    UtcSpeciesComparison:
      properties:
        species_name:
          type: string
          title: Species Name
        mse:
          type: object
          title: Mse
        proximity:
          type: object
          title: Proximity
        output_data:
          anyOf:
          - type: object
          - type: 'null'
          title: Output Data
      type: object
      required:
      - species_name
      - mse
      - proximity
      title: UtcSpeciesComparison
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError